requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.27.1
async-timeout-5.0.1 
asyncpg-0.30.0